+++ src/WiFiManager_Generic_Lite_DUE.h
@@ -8,7 +8,8 @@
 
   Built by Khoi Hoang https://github.com/khoih-prog/WiFiManager_Generic_Lite
   Licensed under MIT license
-  Version: 1.6.0
+  
+  Version: 1.7.0
 
   Version Modified By   Date        Comments
   ------- -----------  ----------   -----------
@@ -18,6 +19,7 @@
   1.5.1   K Hoang      26/01/2022  Update to be compatible with new FlashStorage libraries. Add support to more SAMD/STM32 boards
   1.6.0   K Hoang      26/01/2022  Optional Board_Name in Menu. Optimize code by using passing by reference
                                    Add optional CONFIG_MODE_LED. Add function isConfigMode()
+  1.7.0   K Hoang      27/04/2022  Use WiFiMulti_Generic library for auto-checking / auto-reconnecting MultiWiFi 
  ********************************************************************************************************************************/
 
 #ifndef WiFiManager_Generic_Lite_DUE_h
@@ -34,18 +36,21 @@
 #endif
 
 #ifndef WIFI_MANAGER_GENERIC_LITE_VERSION
-  #define WIFI_MANAGER_GENERIC_LITE_VERSION            "WiFiManager_Generic_Lite v1.6.0"
+  #define WIFI_MANAGER_GENERIC_LITE_VERSION            "WiFiManager_Generic_Lite v1.7.0"
 
   #define WIFI_MANAGER_GENERIC_LITE_VERSION_MAJOR      1
-  #define WIFI_MANAGER_GENERIC_LITE_VERSION_MINOR      6
+  #define WIFI_MANAGER_GENERIC_LITE_VERSION_MINOR      7
   #define WIFI_MANAGER_GENERIC_LITE_VERSION_PATCH      0
 
-#define WIFI_MANAGER_GENERIC_LITE_VERSION_INT          1006000
+#define WIFI_MANAGER_GENERIC_LITE_VERSION_INT          1007000
 
 #endif
 
 #if (USE_WIFI_NINA || USE_WIFI101)
+  #include <WiFiMulti_Generic.h>
   #include <WiFiWebServer.h>
+
+  WiFiMulti_Generic wifiMulti;
 #else
   #warning You have to include another WiFiWebServer, such as ESP8266_AT_WebServer.
 #endif
@@ -400,6 +405,11 @@ class WiFiManager_Generic_Lite
       if (hadConfigData && noConfigPortal && (!isForcedConfigPortal) )
       {
         hadConfigData = true;
+
+#if (USE_WIFI_NINA || USE_WIFI101)        
+        wifiMulti.addAP(WIFI_GENERIC_config.WiFi_Creds[0].wifi_ssid, WIFI_GENERIC_config.WiFi_Creds[0].wifi_pw);
+  			wifiMulti.addAP(WIFI_GENERIC_config.WiFi_Creds[1].wifi_ssid, WIFI_GENERIC_config.WiFi_Creds[1].wifi_pw);
+#endif
 
         if (connectMultiWiFi(RETRY_TIMES_CONNECT_WIFI))
         {
@@ -1313,7 +1323,12 @@ class WiFiManager_Generic_Lite
 
 #if USE_DYNAMIC_PARAMETERS      
       dueFlashStorage_putDynamicData();
-#endif      
+#endif
+
+#if (USE_WIFI_NINA || USE_WIFI101)        
+        wifiMulti.addAP(WIFI_GENERIC_config.WiFi_Creds[0].wifi_ssid, WIFI_GENERIC_config.WiFi_Creds[0].wifi_pw);
+  			wifiMulti.addAP(WIFI_GENERIC_config.WiFi_Creds[1].wifi_ssid, WIFI_GENERIC_config.WiFi_Creds[1].wifi_pw);
+#endif
     }
     
     //////////////////////////////////////////////
@@ -1470,6 +1485,64 @@ class WiFiManager_Generic_Lite
     
     //////////////////////////////////////////////
     
+#if (USE_WIFI_NINA || USE_WIFI101)  
+    
+    bool connectMultiWiFi(int retry_time)
+    {
+			// For general board, this better be 1000 to enable connect the 1st time
+			#define WIFI_MULTI_1ST_CONNECT_WAITING_MS             1000L
+
+			#define WIFI_MULTI_CONNECT_WAITING_MS                   500L
+
+			WG_LOGDEBUG("No WiFi. Trying to scan and reconnect");
+
+			WiFi.disconnect();
+
+			int i = 0;
+
+			uint8_t status = wifiMulti.run();
+
+			delay(WIFI_MULTI_1ST_CONNECT_WAITING_MS);
+
+			while ( ( i++ < (retry_time * 5) ) && ( status != WL_CONNECTED ) )
+			{
+				status = WiFi.status();
+
+				if ( status == WL_CONNECTED )
+				  break;
+				else
+				  delay(WIFI_MULTI_CONNECT_WAITING_MS);
+			}
+
+			if ( status == WL_CONNECTED )
+			{
+				WG_LOGERROR1(F("WiFi connected after time: "), i);
+				WG_LOGERROR3(F("SSID:"), WiFi.SSID(), F(",RSSI="), WiFi.RSSI());
+
+		#if (defined(ESP32) || defined(ESP8266))
+				WG_LOGERROR3(F("Channel:"), WiFi.channel(), F(",IP address:"), WiFi.localIP() );
+		#else
+				WG_LOGERROR1(F("IP address:"), WiFi.localIP() );
+		#endif
+			}
+			else
+			{
+				WG_LOGERROR(F("WiFi not connected"));
+
+				if (wifiMulti.run() != WL_CONNECTED)
+				{
+				  Serial.println("WiFi not connected!");
+				  delay(1000);
+				}
+			}
+
+			return (status == WL_CONNECTED);
+    }
+
+#else
+    
+    //////////////////////////////////////////////
+
 // Max times to try WiFi per loop() iteration. To avoid blocking issue in loop()
 // Default 1 and minimum 1.
 #if !defined(MAX_NUM_WIFI_RECON_TRIES_PER_LOOP)      
@@ -1606,7 +1679,9 @@ class WiFiManager_Generic_Lite
 
       return wifi_connected;  
     }
-    
+
+#endif
+   
     //////////////////////////////////////////////
     
     // NEW